requirements.txt
openai
pandas
faiss-cpu
sentence-transformers
streamlit